tr-metamon/hello_world/requirements.txt
requests
mysql-connector-python==8.2.0